requirements.txt
pandas
numpy
pynacl